Moved tabbed onboarding components to a sub dir
Refs #2025
