+++ ui/pages/Tab.tsx
@@ -4,14 +4,14 @@ import { HashRouter, Route, Switch } from "react-router-dom"
 import { Store } from "webext-redux"
 import BrowserTabContainer from "../components/BrowserTab/BrowserTabContainer"
 import Ledger from "./Ledger/Ledger"
-import TabbedOnboardingAddWallet from "./Onboarding/TabbedOnboardingAddWallet"
-import TabbedOnboardingDone from "./Onboarding/TabbedOnboardingDone"
-import TabbedOnboardingImportSeed from "./Onboarding/TabbedOnboardingImportSeed"
-import TabbedOnboardingSetPassword from "./Onboarding/TabbedOnboardingSetPassword"
-import TabbedOnboardingSaveSeed from "./Onboarding/TabbedOnboardingSaveSeed"
-import TabbedOnboardingVerifySeed from "./Onboarding/TabbedOnboardingVerifySeed"
-import TabbedOnboardingInfoIntro from "./Onboarding/TabbedOnboardingInfoIntro"
-import TabbedOnboardingViewOnlyWallet from "./Onboarding/TabbedOnboardingViewOnlyWallet"
+import TabbedOnboardingAddWallet from "./Onboarding/Tabbed/AddWallet"
+import TabbedOnboardingDone from "./Onboarding/Tabbed/Done"
+import TabbedOnboardingImportSeed from "./Onboarding/Tabbed/ImportSeed"
+import TabbedOnboardingSetPassword from "./Onboarding/Tabbed/SetPassword"
+import TabbedOnboardingSaveSeed from "./Onboarding/Tabbed/SaveSeed"
+import TabbedOnboardingVerifySeed from "./Onboarding/Tabbed/VerifySeed"
+import TabbedOnboardingInfoIntro from "./Onboarding/Tabbed/Intro"
+import TabbedOnboardingViewOnlyWallet from "./Onboarding/Tabbed/ViewOnlyWallet"
 import TabNotFound from "./TabNotFound"
 
 /**
